ChatbotPortal/requirements.txt
certifi==2019.9.11
Django==2.2.6
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
Faker==2.0.2
Markdown==3.1.1
mock==3.0.5
Pillow==6.2.0
python-dateutil==2.8.0
pytz==2019.3
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.6
wincertstore==0.2
